Load Date Picking pattern data from disk based on dynamic route slug
