+++ polaris.shopify.com/content/patterns/date-picking.md
@@ -5,4 +5,6 @@ url: /patterns/date-picking
 previewImg: /images/patterns/pattern-thumbnail-date-picking.png
 status: {value: 'Beta', message: ''}
 order: 10
+githubDiscussionsLink: https://github.com/Shopify/polaris/discussions/7852
+contentFile: 'date-picking.ts'
 ---
