--- polaris.shopify.com/pages/patterns/date-picking.tsx
@@ -1,4 +1,0 @@
-import PatternsDatePickingPage from '../../src/components/PatternsDatePickingPage';
-export default function DatePickingPage() {
-  return <PatternsDatePickingPage />;
-}
